Fix assignment repetition; regenerate tasks
In the previous set of tasks, there was a bug
that caused each task to repeat questions
after a certain point if the number of labelers
evenly divided the number of pairs. This bug
has been addressed, task text files have been
regenerated, and various sanity checks have
been added in the code to ensure this does not
happen again.