Generate graphs for real and semi-synthetic data
Labels from the lab have been processed, and the
generated graphs are saved in `data/graphs` in
the form of a list of edges and corresponding
weights.

Found the distribution and parameters that best
fit the data and used that to create a function
that generates random edge weights for the semi-
synthetic approach.

Moved existing plots to `plots` directory and
generated plots with new edge weight
distribution.